+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 # 1.4.3 (pre-release)
+* [Fix #107] Moving a docked port causes physics effects.
 * [Fix #158] NRE when transferring crew into a pod part without KIS inventories.
 * [Fix #199] Wrong kerbal class in inventory window info when in command seat.
 * [Fix #203] Some physicsless part trigger physics effect on move&attach.